Make Artifactory plugin be the last one in the BuildWrappers list.
In order to guarantee the EnvInject plugin is before this plugin
in the sorted order of the BuildWrappers the @Extension annotation
was changed to use ordinal -2000. The EnvInject plugin was also
modified to use ordinal -1 just so it is the first in the group
od negative ordinals.

(cherry picked from commit 88bfc36c9f608733793e102754badb55c7366e1b)