Replace GUAVA and Docker with versions that avoid incompatibilities
GUAVA is forced in the core WAR file to 27.1-jre by the OWASP formatter
plugin. This has caused issrus in the past and upgrading it here makes
the most sense.

Upgrade Docker API to the same version as what we use in Jenkins core.
This should be migrated to a dependency to the Docker API plugin
instead.